refactoring: 퍼렐럴 라우트 구조 변경, 채팅 룸 이동시  중첩 모달 문제로 인터셉팅 라우트 삭제
